requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.9 